Change convention on 2d coeff naming.
We name the four coefficients of single level 2d FWT
ll, lh, hl, hh based on the filter combination they are
derived from (l - lowpass, h - highpass).
The approximation coeff is clearly denoted by ll,
the diagonal detail coeff by hh.

We generally abbreviate by the axis order, i.e.
lh corresponds to low pass on the first axis and
highpass on the second.

The naming of the horizontal and vertical detail coeffs
depend on the convention how the axes are interpreted.
PyWavelets interpret the first axis as rows and the second
as columns -> horizontal details are hl.